+++ sel/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## sel-0.0.2.0 -- XXXX-XX-XX
 
 * Add usages of `secureMain` in examples
+* Add utility functions and instances to Sel.PublicKey.Signature ([#153](https://github.com/haskell-cryptography/libsodium-bindings/pull/153))